api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.34
psycopg[binary]==3.2.1
redis==5.0.8
rq==1.16.2
pyyaml==6.0.2
alembic==1.13.2
authlib==1.3.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
python-multipart==0.0.9
email-validator==2.2.0
itsdangerous==2.2.0
httpx==0.27.2
